requirements.txt
aiopg==0.10.0
appnope==0.1.0
decorator==4.0.10
ipdb==0.10.1
ipython==5.0.0
ipython-genutils==0.1.0
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.3
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.5.3
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.0.14
traitlets==4.2.2
wcwidth==0.1.7
wheel==0.24.0
